chore: move tests to right file
